Don't add jacoco tasks for ignored build variants
When using a variantFilter to ignore certain variants the plugin was still generating the corresponding tasks

    variantFilter { variant ->
        if (variant.name == "debug") {
            variant.setIgnore(true)
        }
    }

Caused: Could not determine the dependencies of task ':app:jacocoTestReportDebug'.

This broke the gradle `build` command

The new implementation uses android.libraryVariants or android.applicationVariants instead of iterating manually over buildTypes and flavors. The bug is fixed because ignored variants are not included.